DatabaseRepository fix

+++ DatabaseClasses/Activity.cs
@@ -12,7 +12,7 @@ namespace SailingBackend.DatabaseClasses
         public string ActivityName { get; set; }
         public string ActivityDesc { get; set; }
         public string ActivityType { get; set; }
-        public double EntranceFee { get; set; }
+        public float EntranceFee { get; set; }
         public string City { get; set; }
         public string Address { get; set; }
         public float Lat { get; set; }
